pyproject.toml
[project]
name = "research-paper"
version = "0.1.0"
description = ""
authors = [
    {name = "Damian Lebiedź"}
]
license = {text = "Proprietary"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas (>=2.3.3,<2.4.0)",
    "python-binance (>=1.0.30,<1.1.0)",
    "tqdm (>=4.67.1,<4.68.0)",
    "numpy (>=2.3.4,<2.4.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "matplotlib (>=3.10.7,<3.11.0)",
    "statsmodels (>=0.14.5,<0.15.0)",
    "scikit-optimize (>=0.10.2,<0.11.0)",
    "scipy (>=1.16.2,<1.17.0)",
    "joblib (>=1.5.2,<1.6.0)",
    "gymnasium (>=1.2.2,<2.0.0)",
    "stable-baselines3 (>=2.7.1,<3.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "wandb (>=0.23.1,<0.24.0)",
    "pyarrow (>=22.0.0,<23.0.0)"
]


[tool.poetry]
packages = [{ include = "modules" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"